move vectorise_over_dict_args to jax fixtures
